% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{cluster_map}
\alias{cluster_map}
\title{Produces an interactive map to visualize the results of an analysis}
\usage{
cluster_map(results, significant_only = TRUE, alpha = 0.05)
}
\arguments{
\item{results}{data.table containing the formatted results of the analysis.}

\item{significant_only}{Boolean value. If true, only map the significant clusters.}

\item{alpha}{Specifies the significance level for which to display clusters.}
}
\description{
Produces an interactive map to visualize the results of an analysis
}
