% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_dist}
\alias{max_dist}
\title{Calculates the max distance between points within a cluster. This corresponds to the radius of the cluster}
\usage{
max_dist(idlist, idvec, locs)
}
\arguments{
\item{idlist}{A list of vectors containing the ID's in each cluster}

\item{idvec}{A vector of ids that are associated with each observation}

\item{locs}{A matrix of coordinates for each observation/centroid}
}
\description{
Calculates the max distance between points within a cluster. This corresponds to the radius of the cluster
}
