% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multinom_mlc}
\alias{multinom_mlc}
\title{Calculates the most likely cluster for each centroid point by calculating the test statistic for all scanning windows}
\usage{
multinom_mlc(zones, group, id, levels)
}
\arguments{
\item{zones}{list of space-time scanning windows}

\item{group}{vector corresponding to group membership of each observation}

\item{id}{vector corresponding to the ID number of the observation}

\item{levels}{vector denoting the unique levels of the group variable to which an observation can belong}
}
\description{
Calculates the most likely cluster for each centroid point by calculating the test statistic for all scanning windows
}
