% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multinom_permutation}
\alias{multinom_permutation}
\title{Conducts the permutations needed for Monte Carlo hypothesis testing}
\usage{
multinom_permutation(zones, group, levels, n_perm)
}
\arguments{
\item{zones}{list of space-time scanning windows}

\item{group}{vector corresponding to group membership of each observation}

\item{n_perm}{number of Monte Carlo permutations to perform}
}
\description{
Conducts the permutations needed for Monte Carlo hypothesis testing
}
