\name{multinomss-package}
\alias{multinomss-package}
\alias{multinomss}
\docType{package}
\title{\packageTitle{multinomss}}
\description{\packageDescription{multinomss}}
\section{Package Content}{\packageIndices{multinomss}}
\author{\packageAuthor{multinomss}}
\section{Maintainer}{\packageMaintainer{multinomss}}
\keyword{package}
