% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{non_overlap}
\alias{non_overlap}
\title{Gives the indices of the non-overlapping scanning windows.}
\usage{
non_overlap(ids)
}
\arguments{
\item{ids}{A list of vectors containing the ID numbers associated with the observations in each cluster}
}
\description{
Gives the indices of the non-overlapping scanning windows.
}
