% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_val}
\alias{p_val}
\title{Calculates the p-values from the Monte Carlo permutation-generated test statistics}
\usage{
p_val(stats, perm)
}
\arguments{
\item{stats}{vector of test statistics from the data for which p-values are desired}

\item{perm}{vector of permutation test statistics used to compute p-values}
}
\description{
Calculates the p-values from the Monte Carlo permutation-generated test statistics
}
