% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{results}
\alias{results}
\title{Used to format results for output}
\usage{
results(mlcs, perm_p, ids)
}
\arguments{
\item{mlcs}{A list produced by the \code{multinom_mlc} function}

\item{perm_p}{Permutation p-values produced by the \code{multinom_permutation} function}

\item{ids}{A vector that corresponds to the ID of each observation}
}
\description{
Used to format results for output
}
