% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{time_intervals}
\alias{time_intervals}
\title{Used to aggregate time by days, weeks, months or years}
\usage{
time_intervals(
  date,
  agg_unit = c("days", "weeks", "months", "years"),
  agg_interval
)
}
\description{
@param date A vector of dates to be aggregated
 @param agg_unit Determines what time unit to aggregate across
 @param agg_interval Determines how many time units to aggregate across 
 @export
}
