% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{cluster_map}
\alias{cluster_map}
\title{Produces an interactive map to visualize the results of an analysis}
\usage{
cluster_map(results, mlc_only = TRUE, map_layer = "CartoDB.Voyager")
}
\arguments{
\item{results}{data.table containing the formatted results of the analysis.}

\item{mlc_only}{Boolean value. If true, only map the most likely cluster.}

\item{map_layer}{Specifies the desired base tile layer for the map. Available layers can be previewed at https://leaflet-extras.github.io/leaflet-providers/preview/.}
}
\description{
Produces an interactive map to visualize the results of an analysis
}
