% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatial_zones}
\alias{spatial_zones}
\title{Calculates the spatial zones, starting with each centroid and increasing size
from 1 to pre-specified k points. These represent circular zones in space.}
\usage{
spatial_zones(locs, k)
}
\arguments{
\item{locs}{An n x 2 matrix of centroid point coordinates in latitude/longitude}

\item{k}{An integer specifying the maximum number of nearest neighbors to include in spatial zones}
}
\description{
Calculates the spatial zones, starting with each centroid and increasing size
from 1 to pre-specified k points. These represent circular zones in space.
}
