% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{temporal_zones}
\alias{temporal_zones}
\title{Takes spatial zone input and calculates the temporal zones. For each circular
zone in space, it effectively varies the height to create cylindrical zones.}
\usage{
temporal_zones(zones, dates, time_prop, study_length)
}
\arguments{
\item{zones}{A list of spatial zones produced by \code{spatial_zones()} function}

\item{dates}{A vector containing the dates associated with each observation. If time aggregation is used, this will correspond to time interval that each observation falls into}

\item{time_prop}{A float/decimal specifying the maximum proportion of the study length to consider as a cluster}

\item{study_length}{An integer specifying the length of the study in days}
}
\description{
Takes spatial zone input and calculates the temporal zones. For each circular
zone in space, it effectively varies the height to create cylindrical zones.
}
